requirements1.txt
absl-py==2.1.0
accelerate==1.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.8
aiosignal==1.3.1
albucore==0.0.17
albumentations==1.4.18
alembic==1.14.0
annotated-types==0.7.0
anyio==4.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
breadability==0.1.20
Brotli==1.1.0
cachetools==5.5.2
certifi==2024.6.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colorlog==6.9.0
colorlover==0.3.0
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
Django==5.1.2
docker-pycreds==0.4.0
docopt==0.6.2
eval_type_backport==0.2.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
fonttools==4.54.1
fpdf==1.7.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
git-filter-repo==2.45.0
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.69.2
googlesearch-python==1.2.4
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.25.1
idna==3.7
imageio==2.36.0
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.11
kaggle==1.6.17
keras==3.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
kt-legacy==1.0.5
lazy_loader==0.4
libclang==18.1.1
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mutagen==1.47.0
namex==0.0.8
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.12.1
optuna==4.0.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
plotly==5.24.1
portalocker==2.10.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.26.1
protobuf==5.29.4
psutil==6.0.0
pure-eval==0.2.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pyphen==0.17.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.2
#pywin32==310
#pywinpty==2.0.15
PyYAML==6.0.1
pyzmq==26.0.3
QtPy==2.4.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.8.1
rouge_score==0.1.2
rpds-py==0.18.1
rsa==4.9
sacrebleu==2.4.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.3
setuptools==70.1.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.36
sqlparse==0.5.1
stack-data==0.6.3
sumy==0.11.0
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.18.0.post0
textstat==0.7.4
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.3.0
tokenizers==0.20.0
torch==2.4.1
torchaudio==2.4.1
torchsummary==1.5.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
trio==0.26.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.2
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
waitress==3.0.2
wandb==0.18.6
wcwidth==0.2.13
webcolors==24.6.0
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==13.1
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.11
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.5.0
yarl==1.13.1
yt-dlp==2024.10.7
